move authtype logging to api/auth handler
