+++ handler/api/auth/auth_test.go
@@ -86,3 +86,21 @@ func TestAuth_Guest(t *testing.T) {
 		t.Errorf("Want status code %d, got %d", want, got)
 	}
 }
+
+func TestAuthType(t *testing.T) {
+	cookieRequest := httptest.NewRequest("GET", "/", nil)
+	if authType(cookieRequest) != "cookie" {
+		t.Error("authtype is not cookie")
+	}
+
+	headerRequest := httptest.NewRequest("GET", "/", nil)
+	headerRequest.Header.Add("Authorization", "test")
+	if authType(headerRequest) != "token" {
+		t.Error("authtype is not token")
+	}
+
+	formRequest := httptest.NewRequest("GET", "/?access_token=test", nil)
+	if authType(formRequest) != "token" {
+		t.Error("authtype is not token")
+	}
+}
