+++ logger/handler_test.go
@@ -7,7 +7,6 @@
 package logger
 
 import (
-	"net/http/httptest"
 	"testing"
 )
 
@@ -17,22 +16,4 @@ func TestMiddleware(t *testing.T) {
 
 func TestMiddleware_GenerateRequestID(t *testing.T) {
 	t.Skip()
-}
-
-func TestAuthType(t *testing.T) {
-	cookieRequest := httptest.NewRequest("GET", "/", nil)
-	if authType(cookieRequest) != "cookie" {
-		t.Error("authtype is not cookie")
-	}
-
-	headerRequest := httptest.NewRequest("GET", "/", nil)
-	headerRequest.Header.Add("Authorization", "test")
-	if authType(headerRequest) != "token" {
-		t.Error("authtype is not token")
-	}
-
-	formRequest := httptest.NewRequest("GET", "/?access_token=test", nil)
-	if authType(formRequest) != "token" {
-		t.Error("authtype is not token")
-	}
 }
